refactor: mv isFormatOf handling up
so that MediaMetadataList can be used multiple times
